test/web-platform-tests/to-upstream/html/semantics/document-metadata/the-style-element/sheet-with-empty-style.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>The sheet property should work even with an empty style element</title>
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<link rel="help" href="https://html.spec.whatwg.org/multipage/semantics.html#update-a-style-block">
<link rel="help" href="https://drafts.csswg.org/cssom/#the-linkstyle-interface">

<style></style>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

setup({ single_test: true });

assert_not_equals(document.querySelector("style").sheet, null);
assert_true(document.querySelector("style").sheet instanceof CSSStyleSheet);

done();
</script>
